new media api server

+++ media-api/src/api/models/tagModel.ts
@@ -1,16 +1,13 @@
 import {ResultSetHeader, RowDataPacket} from 'mysql2';
-import {TagResult} from '@sharedTypes/DBTypes';
+import {MediaItem, Tag, TagResult} from '@sharedTypes/DBTypes';
 import promisePool from '../../lib/db';
-import {fetchData} from '../../lib/functions';
 import {MessageResponse} from '@sharedTypes/MessageTypes';
 
 // Request a list of tags
-const fetchAllTags = async (): Promise<TagResult[] | null> => {
+const fetchAllTags = async (): Promise<Tag[] | null> => {
   try {
-    const [rows] = await promisePool.execute<RowDataPacket[] & TagResult[]>(
-      `SELECT Tags.tag_id, Tags.tag_name, MediaItemTags.media_id
-       FROM Tags
-       JOIN MediaItemTags ON Tags.tag_id = MediaItemTags.tag_id`
+    const [rows] = await promisePool.execute<RowDataPacket[] & Tag[]>(
+      'SELECT * FROM Tags'
     );
     if (rows.length === 0) {
       return null;
@@ -22,58 +19,65 @@ const fetchAllTags = async (): Promise<TagResult[] | null> => {
   }
 };
 
-// Post a new tag
-const postTag = async (
-  tag: Omit<TagResult, 'tag_id'>
-): Promise<MessageResponse | null> => {
-  const connection = await promisePool.getConnection();
+const fetchFilesByTagById = async (
+  tag_id: number
+): Promise<MediaItem[] | null> => {
   try {
-    await connection.beginTransaction();
-    const [tagResult] = await connection.execute<ResultSetHeader>(
-      'INSERT INTO Tags (tag_name) VALUES (?)',
-      [tag.tag_name]
-    );
-    if (tagResult.affectedRows === 0) {
-      return null;
-    }
-
-    const [mediaItemTagResult] = await connection.execute<ResultSetHeader>(
-      'INSERT INTO MediaItemTags (media_id, tag_id) VALUES (?, ?)',
-      [tag.media_id, tagResult.insertId]
-    );
-
-    await connection.commit();
-
-    if (mediaItemTagResult.affectedRows === 0) {
-      return null;
-    }
-
-    return {message: 'Tag created'};
-  } catch (e) {
-    await connection.rollback();
-    console.error('postTag error', (e as Error).message);
-    throw new Error((e as Error).message);
-  } finally {
-    connection.release();
-  }
-};
-
-// Request a list of media items by tag
-const fetchMediaByTag = async (tag: string): Promise<TagResult[] | null> => {
-  try {
-    const [rows] = await promisePool.execute<RowDataPacket[] & TagResult[]>(
-      `SELECT Tags.tag_id, Tags.tag_name, MediaItemTags.media_id
-       FROM Tags
-       JOIN MediaItemTags ON Tags.tag_id = MediaItemTags.tag_id
-       WHERE Tags.tag_name = ?`,
-      [tag]
+    console.log(tag_id);
+    const [rows] = await promisePool.execute<RowDataPacket[] & MediaItem[]>(
+      `SELECT * FROM MediaItems
+       JOIN MediaItemTags ON MediaItems.media_id = MediaItemTags.media_id
+       WHERE MediaItemTags.tag_id = ?`,
+      [tag_id]
     );
     if (rows.length === 0) {
       return null;
     }
     return rows;
   } catch (e) {
-    console.error('fetchMediaByTag error', (e as Error).message);
+    console.error('fetchTagByName error', (e as Error).message);
+    throw new Error((e as Error).message);
+  }
+};
+
+// Post a new tag
+const postTag = async (
+  tag_name: string,
+  media_id: number
+): Promise<MessageResponse | null> => {
+  try {
+    let tag_id = 0;
+    // check if tag exists (case insensitive)
+    const [tagResult] = await promisePool.query<RowDataPacket[] & Tag[]>(
+      'SELECT tag_id FROM Tags WHERE tag_name = ?',
+      [tag_name]
+    );
+    if (tagResult.length === 0) {
+      // if tag does not exist create it
+      const [insertResult] = await promisePool.execute<ResultSetHeader>(
+        'INSERT INTO Tags (tag_name) VALUES (?)',
+        [tag_name]
+      );
+      if (insertResult.affectedRows === 0) {
+        return null;
+      }
+      // get tag_id from created tag
+      tag_id = insertResult.insertId;
+    } else {
+      // if tag exists get tag_id from the first result
+      tag_id = tagResult[0].tag_id;
+    }
+    const [MediaItemTagsResult] = await promisePool.execute<ResultSetHeader>(
+      'INSERT INTO MediaItemTags (tag_id, media_id) VALUES (?, ?)',
+      [tag_id, media_id]
+    );
+    if (MediaItemTagsResult.affectedRows === 0) {
+      return null;
+    }
+
+    return {message: 'Tag added'};
+  } catch (e) {
+    console.error('postTagToMedia error', (e as Error).message);
     throw new Error((e as Error).message);
   }
 };
@@ -103,18 +107,24 @@ const deleteTag = async (id: number): Promise<MessageResponse | null> => {
   const connection = await promisePool.getConnection();
   try {
     await connection.beginTransaction();
-    const [tagResult] = await connection.execute<ResultSetHeader>(
-      'DELETE FROM Tags WHERE tag_id = ?',
-      [id]
-    );
-    if (tagResult.affectedRows === 0) {
-      return null;
-    }
 
     const [mediaItemTagResult] = await connection.execute<ResultSetHeader>(
       'DELETE FROM MediaItemTags WHERE tag_id = ?',
       [id]
     );
+
+    if (mediaItemTagResult.affectedRows === 0) {
+      return null;
+    }
+
+    const [tagResult] = await connection.execute<ResultSetHeader>(
+      'DELETE FROM Tags WHERE tag_id = ?',
+      [id]
+    );
+
+    if (tagResult.affectedRows === 0) {
+      return null;
+    }
 
     await connection.commit();
 
@@ -132,4 +142,42 @@ const deleteTag = async (id: number): Promise<MessageResponse | null> => {
   }
 };
 
-export {fetchAllTags, postTag, fetchMediaByTag, fetchTagsByMediaId, deleteTag};
+const deleteTagFromMedia = async (
+  tag_id: number,
+  media_id: number,
+  user_id: number
+): Promise<MessageResponse | null> => {
+  try {
+    // check if user owns media item
+    const [mediaItem] = await promisePool.execute<RowDataPacket[]>(
+      'SELECT * FROM MediaItems WHERE media_id = ? AND user_id = ?',
+      [media_id, user_id]
+    );
+
+    if (mediaItem.length === 0) {
+      throw new Error('Media item not found or user does not own media item');
+    }
+
+    const [result] = await promisePool.execute<ResultSetHeader>(
+      'DELETE FROM MediaItemTags WHERE tag_id = ? AND media_id = ?',
+      [tag_id, media_id]
+    );
+    if (result.affectedRows === 0) {
+      return null;
+    }
+
+    return {message: 'Tag deleted from media item'};
+  } catch (e) {
+    console.error('deleteTagFromMedia error', (e as Error).message);
+    throw new Error((e as Error).message);
+  }
+};
+
+export {
+  fetchAllTags,
+  postTag,
+  fetchTagsByMediaId,
+  fetchFilesByTagById,
+  deleteTag,
+  deleteTagFromMedia,
+};
